Fix automated safety orders (#370)
* Fix automated safety orders

* Fix edit bot update deal

* Fix safety orders autotrade pricing

Reverted to static safety orders. Dynamic does not work well as a safety
net

* Add research test dynamic take_profit and trailling

* Add missing paper trading switch

* Improve and fix deal updates

* Fix excessive 502s

* Do not trigger trailling_stop_loss below buy_price

Doing so will cause a loss.

* Remove unused charting files

+++ web/src/components/BotInfo.jsx
@@ -48,7 +48,7 @@ export default function BotInfo({ bot }) {
                     <th scope="row">{deal.order_id}</th>
                     <td>{deal.deal_type}</td>
                     <td>{deal.price}</td>
-                    <td>{parseInt(deal.qty)}</td>
+                    <td>{parseFloat(deal.qty)}</td>
                     <td>{deal.status}</td>
                     <td>{deal.order_side}</td>
                   </tr>
